Added integration tests for endorsement service
